Add 'data' directory to repository. This is referenced in the retrieval scripts and would likely throw errors if not present.
Ultimately, the data directory would be defined in config.json, but this should hopefully get it working. Didn't add to the repo before becase, well, it's data, not code!

Had to create this here after a week of running since 24 JSON files per day was making it hard to find the programme files.

See #1