Tests out OK now.
Updated get-followers-ids.php to the version I'm using - the Twitter handles to follow are now defined in watchlist.php,

I've also added a variable in there $verbose true|false. Set to true for testing, set to false if running as a cron job and all is well. Each call (takes 3 to pick up the RS followers IDs) is logged to screen so that you can see something's actually happening.

See #1